Flutter app is in a stable point for deployment to android device (#51)
* Updated the launch icon

* some android support

* Seems to be an issue with the datatable, this will do for now

* Trying to fix a bug

* Fixed the Json parsing for the data and recommendations

* fixed the weird _Smi to double issue.

+++ Apps/gm-flutter-app/gimmillions/lib/models/stock-data.dart
@@ -20,7 +20,15 @@ class StockData {
   }
 
   factory StockData.fromJson(Map<String, dynamic> json) {
-    return StockData(DateTime.parse(json['date']), json['symbol'], json['open'], json['high'], json['low'],
-        json['close'], json['adjustedClose'], json['volume'] as double, json['previousClose']);
+    return StockData(
+        DateTime.parse(json['date']),
+        json['symbol'],
+        json['open'].toDouble(),
+        json['high'].toDouble(),
+        json['low'].toDouble(),
+        json['close'].toDouble(),
+        json['adjustedClose'].toDouble(),
+        json['volume'].toDouble(),
+        json['previousClose'].toDouble());
   }
 }
